+++ src/dsp.zig
@@ -18,6 +18,7 @@ pub const Limiter = @import("dsp/limiter.zig").Limiter;
 pub const Filter = @import("dsp/filter.zig").Filter;
 pub const Utility = @import("dsp/utility.zig").Utility;
 pub const StereoWidth = @import("dsp/stereo_width.zig").StereoWidth;
+pub const AutoPan = @import("dsp/auto_pan.zig").AutoPan;
 pub const Gate = @import("dsp/gate.zig").Gate;
 pub const Saturator = @import("dsp/saturator.zig").Saturator;
 pub const Crusher = @import("dsp/crusher.zig").Crusher;
@@ -72,6 +73,7 @@ test {
     _ = Limiter;
     _ = Utility;
     _ = StereoWidth;
+    _ = AutoPan;
     _ = Gate;
     _ = Saturator;
     _ = Crusher;